requirements-08022020.txt
alabaster==0.7.12
amqp==2.5.1
appdirs==1.4.3
attrs==19.2.0
autoflake==1.3.1
autopep8==1.4.4
Babel==2.7.0
backports.csv==1.0.7
billiard==3.6.1.0
black==19.3b0
bottle==0.12.17
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-appconf==1.0.3
django-avatar==4.1.0
django-axes==5.0.18
django-bootstrap-toolkit==2.15.0
django-bootstrap3==11.1.0
django-braces==1.13.0
django-celery-beat==1.5.0
django-compressor==2.3
django-crispy-forms==1.8.0
django-datatables-view==1.19.1
django-debug-toolbar==2.0
django-excel==0.0.10
django-filebrowser==3.11.1
django-filter==2.2.0
django-formtools==2.1
django-geojson==2.12.0
django-grappelli==2.12.3
django-import-export==1.2.0
django-ipware==2.1.0
django-json-tag==0.0.1
django-leaflet==0.24.0
django-maploom==1.7.0
django-material==1.6.0
django-simple-history==2.7.3
django-timezone-field==3.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-datatables==0.5.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.8
future==0.18.2
Glances==3.1.3
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
jdcal==1.4.1
Jinja2==2.10.3
kombu==4.6.5
lml==0.0.9
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==2.5.14
packaging==19.2
pi==0.1.2
Pillow==6.1.0
psutil==5.6.5
psycopg2==2.8.3
pycodestyle==2.5.0
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
pyflakes==2.1.1
Pygments==2.5.1
pyparsing==2.4.5
python-crontab==2.3.8
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.2
PyYAML==5.1.2
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tablib==0.13.0
texttable==1.6.2
toml==0.10.0
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
